scripts: resync all build scripts

+++ build_kernel.sh
@@ -41,29 +41,26 @@ patch_kernel () {
 	cd "${DIR}/" || exit
 }
 
-flash_kernel_db () {
-	cat ./KERNEL/arch/${KERNEL_ARCH}/boot/dts/*.dts | grep 'model =' | grep -v ',model' | grep -v 'audio' | grep -v 'sgtl5000' | grep -v 'n-board' | awk -F'"' '{print $2}' | sort -u > /tmp/pre.db
-	sed -i -e 's/^/Machine: /' /tmp/pre.db
-	awk '{print $0 "\nMethod: generic\n"}' /tmp/pre.db > patches/all.db
-}
-
 copy_defconfig () {
 	cd "${DIR}/KERNEL" || exit
 	make ARCH=${KERNEL_ARCH} CROSS_COMPILE="${CC}" distclean
 
-	#./ti_config_fragments/defconfig_builder.sh -l
-	./ti_config_fragments/defconfig_builder.sh -t ti_sdk_am3x_release
-	./ti_config_fragments/defconfig_builder.sh -t ti_sdk_dra7x_release
+	if [ -f ./ti_config_fragments/defconfig_builder.sh ] ; then
+		#./ti_config_fragments/defconfig_builder.sh -l
+		./ti_config_fragments/defconfig_builder.sh -t ti_sdk_am3x_release
+		./ti_config_fragments/defconfig_builder.sh -t ti_sdk_dra7x_release
 
-	make ARCH=${KERNEL_ARCH} CROSS_COMPILE="${CC}" distclean
-	make ARCH=${KERNEL_ARCH} CROSS_COMPILE="${CC}" ti_sdk_am3x_release_defconfig
-	cp -v .config "${DIR}/patches/ti_sdk_am3x_release_defconfig"
+		make ARCH=${KERNEL_ARCH} CROSS_COMPILE="${CC}" distclean
+		make ARCH=${KERNEL_ARCH} CROSS_COMPILE="${CC}" ti_sdk_am3x_release_defconfig
+		cp -v .config "${DIR}/patches/ti_sdk_am3x_release_defconfig"
 
-	make ARCH=${KERNEL_ARCH} CROSS_COMPILE="${CC}" distclean
-	make ARCH=${KERNEL_ARCH} CROSS_COMPILE="${CC}" ti_sdk_dra7x_release_defconfig
-	cp -v .config "${DIR}/patches/ti_sdk_dra7x_release_defconfig"
+		make ARCH=${KERNEL_ARCH} CROSS_COMPILE="${CC}" distclean
+		make ARCH=${KERNEL_ARCH} CROSS_COMPILE="${CC}" ti_sdk_dra7x_release_defconfig
+		cp -v .config "${DIR}/patches/ti_sdk_dra7x_release_defconfig"
 
-	make ARCH=${KERNEL_ARCH} CROSS_COMPILE="${CC}" distclean
+		make ARCH=${KERNEL_ARCH} CROSS_COMPILE="${CC}" distclean
+	fi
+
 	if [ ! -f "${DIR}/.yakbuild" ] ; then
 		make ARCH=${KERNEL_ARCH} CROSS_COMPILE="${CC}" "${config}"
 		cp -v .config "${DIR}/patches/ref_${config}"
@@ -228,9 +225,6 @@ if [ "${FULL_REBUILD}" ] ; then
 	fi
 
 	patch_kernel
-	if [ ! "${AUTO_BUILD}" ] ; then
-		flash_kernel_db
-	fi
 	copy_defconfig
 fi
 if [ ! "${AUTO_BUILD}" ] ; then
